test (run): Fix the rest of the integration tests for ExpSteps
